+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of the splunk cookbook.
     - Adds logic to skip any initialization or bootstrapping of search head cluster resources.
 - Fixes issue [#140](https://github.com/chef-cookbooks/chef-splunk/issues/140)
     - Adds back resource actions: :enable, :disable, :install, :remove
+- Integrates a search head cluster to a single or multisite indexer cluster
 
 ## 4.0.5 (2020-01-15)
 - Adds a state file and a guard that ensures the `template[user-seed.conf]` resource is idempotent